#84783 - Integração com a EW

+++ TronBox.Domain/DTO/InnerClassDTO/DadosImportacaoDTO.cs
@@ -2,7 +2,7 @@
 {
     public class DadosImportacaoDTO
     {
-        public int DataImportacao { get; set; }
+        public long DataImportacao { get; set; }
         public string Usuario { get; set; }
     }
 }
